get_head/requirements.txt
torch==1.10.0+cu113
numpy==1.21.1
Pillow==8.4.0
